requirements.txt
pyspark==3.5.1
pyarrow==16.1.0
pandas==2.2.2
apache-airflow==2.8.4
apache-airflow-providers-apache-spark==4.5.0
